+++ R/sim_design.R
@@ -16,6 +16,7 @@
 #' @param seed a single value, interpreted as an integer, or NULL (see set.seed)
 #' @param interactive whether to run the function interactively
 #' @param design a design list including within, between, n, mu, sd, r, dv, id
+#' @param rep the number of data frames to return (default 1); if greater than 1, the returned data frame is nested by rep
 #' 
 #' @return a tbl
 #' 
@@ -29,7 +30,7 @@ sim_design <- function(within = list(), between = list(),
                        id = list(id = "id"),
                        plot = faux_options("plot"), 
                        seed = NULL, interactive = FALSE, 
-                       design = NULL) {
+                       design = NULL, rep = 1) {
   # check the design is specified correctly
   if (interactive) {
     design <- interactive_design(plot = plot)
@@ -47,10 +48,9 @@ sim_design <- function(within = list(), between = list(),
   }
   
   # simulate the data
-  data <- sim_data(design, empirical = empirical, long = long, seed = seed)
+  data <- sim_data(design, empirical = empirical, long = long, seed = seed, rep = rep)
   
   attr(data, "design") <- design
-  class(data) <- c("faux", "data.frame")
   
   return(data)
 }
@@ -60,6 +60,7 @@ sim_design <- function(within = list(), between = list(),
 #' @param design A list of design parameters created by check_design()
 #' @param empirical logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance 
 #' @param long Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format
+#' @param rep the number of data frames to return (default 1); if greater than 1, the returned data frame is nested by rep
 #' @param seed a single value, interpreted as an integer, or NULL (see set.seed)
 #' @param sep separator for within-columns, defaults to _ (see tidyr::separate)
 #' 
@@ -67,7 +68,15 @@ sim_design <- function(within = list(), between = list(),
 #' @export
 #' 
 sim_data <- function(design, empirical = FALSE, long = FALSE, 
-                        seed = NULL, sep = faux_options("sep")) {
+                     rep = 1, seed = NULL, sep = faux_options("sep")) {
+  if (!is.numeric(rep)) {
+    stop("rep must be a number")
+  } else if (rep < 1) {
+    stop("rep must be >= 1")
+  } else if (rep%%1 != 0) {
+    warning("rep should be an integer")
+  }
+  
   if (!is.null(seed)) {
     # reinstate system seed after simulation
     sysSeed <- .GlobalEnv$.Random.seed
@@ -88,8 +97,8 @@ sim_data <- function(design, empirical = FALSE, long = FALSE,
   id <- names(id)
   
   # define columns
-  cells_w <- cell_combos(within, dv)
-  cells_b <- cell_combos(between, dv) 
+  cells_w <- faux:::cell_combos(within, dv)
+  cells_b <- faux:::cell_combos(between, dv) 
   
   # get factor names
   within_factors <- names(within)
@@ -100,29 +109,37 @@ sim_data <- function(design, empirical = FALSE, long = FALSE,
   if (length(within_factors) == 0)  within_factors  <- ".tmpvar."
   
   # simulate data for each between-cell
-  for (cell in cells_b) {
-    cell_vars <- rnorm_multi(
-      n = n[[cell]], 
-      vars = length(cells_w), 
-      mu = mu[[cell]] %>% unlist(), 
-      sd = sd[[cell]] %>% unlist(), 
-      r = r[[cell]], 
-      varnames = cells_w, 
-      empirical = empirical
-    ) %>%
-      dplyr::mutate("btwn" = cell)
-    
-    # add cell values to df
-    if (cell == cells_b[1]) { 
-      df <- cell_vars # first cell sets up the df
+  for (rep_n in 1:rep) {
+    for (cell in cells_b) {
+      cell_vars <- rnorm_multi(
+        n = n[[cell]], 
+        vars = length(cells_w), 
+        mu = mu[[cell]] %>% unlist(), 
+        sd = sd[[cell]] %>% unlist(), 
+        r = r[[cell]], 
+        varnames = cells_w, 
+        empirical = empirical
+      )
+      cell_vars$btwn <- cell
+      
+      # add cell values to df
+      if (cell == cells_b[1]) { 
+        sub_df <- cell_vars # first cell sets up the df
+      } else {
+        sub_df <- dplyr::bind_rows(sub_df, cell_vars)
+      }
+    }
+    sub_df$`.rep.` <- rep_n
+    if (rep_n == 1) {
+      df <- sub_df
     } else {
-      df <- dplyr::bind_rows(df, cell_vars)
+      df <- dplyr::bind_rows(df, sub_df)
     }
   }
   
   # set column order
-  col_order <- c(id, between_factors, cells_w) %>%
-    setdiff(".tmpvar.")
+  col_order <- c(".rep.", id, between_factors, cells_w) %>%
+    setdiff(c(".tmpvar."))
   
   # figure out total number of subjects
   sub_n <- unlist(n) %>% sum()
@@ -130,7 +147,9 @@ sim_data <- function(design, empirical = FALSE, long = FALSE,
   # create wide dataframe
   df_wide <- df %>%
     tidyr::separate("btwn", between_factors, sep = sep) %>%
+    dplyr::group_by(`.rep.`) %>%
     dplyr::mutate(!!id := make_id(sub_n)) %>%
+    dplyr::ungroup(`.rep.`) %>%
     dplyr::mutate_at(c(between_factors), ~as.factor(.)) %>%
     dplyr::select(tidyselect::one_of(col_order))
   
@@ -142,7 +161,7 @@ sim_data <- function(design, empirical = FALSE, long = FALSE,
   
   if (long == TRUE && length(within)) {
     # not necessary for fully between designs
-    col_order <- c(id, between_factors, within_factors, dv) %>%
+    col_order <- c(".rep.", id, between_factors, within_factors, dv) %>%
       setdiff(".tmpvar.")
     
     df_long <- df_wide %>%
@@ -156,8 +175,21 @@ sim_data <- function(design, empirical = FALSE, long = FALSE,
     for (f in factors_to_order) {
       df_long[[f]] <- factor(df_long[[f]], levels = names(within[[f]]))
     }
-    return(df_long)
+
+    df_return <- df_long
+  } else {
+    df_return <- df_wide
   }
   
-  return(df_wide)
+  class(df_return) <- c("faux", "data.frame")
+  
+  if (rep == 1) {
+    df_return$`.rep.` <- NULL
+  } else {
+    df_return <- df_return %>%
+      dplyr::group_by(`.rep.`) %>%
+      tidyr::nest() %>%
+      dplyr::rename("rep" = `.rep.`)
+  }
+  return(df_return)
 }